refactored jet class
