+++ app/src/main/java/hacking/to/the/gate/JetLifeCycle.java
@@ -1,8 +1,10 @@
 package hacking.to.the.gate;
 
+import jet.EnemyJet;
+
 /**
  * Created by Ruiqian on 9/6/2015.
  */
 public interface JetLifeCycle {
-    void onDeath(Jet jet);
+    void onDeath(EnemyJet jet);
 }
